fixed types, changed IGQLCalendarEvent to IHW
