+++ packages/spruce-types/index.ts
@@ -3,4 +3,5 @@ export * from './src/generated/api-gql'
 export { Maybe, Scalars } from './src/combinedGQL'
 export * from './src/events'
 export * from './src/events/reschedule-calendar-event'
+export * from './src/events/get-calendars'
 export * from './src/core'
