+++ packages/spruce-types/src/events.ts
@@ -1,6 +1,3 @@
-import { IHWCalendarEvent } from './generated/hw-gql'
-import { ICoreGQLCalendar } from './generated/api-gql'
-
 export interface ISpruceEventV1User {
 	/** The User's id */
 	id: string
@@ -166,37 +163,4 @@ export interface ISpruceEventWarning {
 	reason: string
 	/** A description of the warning that can be displayed to the user */
 	friendlyReason: string
-}
-
-/**
- * Calendars
- */
-/** Payload sent to your event listener when listening to: get-calendars */
-export interface IEventGetCalendarsPayload {
-	/** the route the user is viewing (like calendar_location) */
-	route: string
-}
-
-/** set to the body of as a response get-calendars event */
-export type IEventGetCalendarsBody = ICoreGQLCalendar[]
-
-/**
- * Calendar Events
- */
-/** Payload sent to your event listener when listening to: get-calendar-events */
-export interface IEventGetCalendarEventsPayload {
-	/** id of calendar we need events for */
-	calendarId: string
-	/** user id's of users visible in the calendar */
-	userIds: string[]
-	/** start date/time in ISO8601 */
-	startAt: string
-	/** end date/time in ISO6801  */
-	endAt: string
-}
-
-/** Body of response to ge-calendar-events event */
-export interface IEventGetCalendarEventsBody {
-	/** An array of calendar events */
-	calendarEvents?: IHWCalendarEvent[]
 }
